requirements/prod.txt
django==4.2
django-cleanup==8.1.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-slugify==8.0.4
sorl-thumbnail==12.10.0